requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
nltk==3.8.1
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2